Tidy up code and prepare for sonde updates
