+++ package.json
@@ -9,6 +9,6 @@
     "mqtt": "^4.3.7"
   },
   "scripts": {
-    "start": "node main.js"
+    "start": "node core.js"
   }
 }
